Made search_anime a bot method
