+++ lib/commands/fun/anime.py
@@ -5,28 +5,7 @@ import gc
 from discord.ext import commands
 from bs4 import BeautifulSoup
 from settings import *
-
-
-class AnimeSearchResult:
-    def __init__(self, title, id, url):
-        self._title = title
-        self._id = id
-        self._url = url
-
-
-    @property
-    def title(self):
-        return self._title
-
-
-    @property
-    def id(self):
-        return self._id
-
-
-    @property
-    def url(self):
-        return self._url
+from load import *
 
 
 @bot.command(
@@ -39,33 +18,23 @@ async def _anime(cmd, *, query):
     if len(query) < 3:
         await cmd.send(f"Search query must have at least 3 characters")
         return
-    rslt = []
-    url = f"https://myanimelist.net/anime.php?q={query}"
-    async with bot.session.get(url) as response:
-        if response.status == 200:
-            html = await response.text()
-            soup = BeautifulSoup(html, "html.parser")
-            obj = soup.find_all(name = "a", attrs = {"class": "hoverinfo_trigger fw-b fl-l"}, limit = 6)
-            for tag in obj:
-                url = tag.get("href")
-                id = int(url.split("/")[4])
-                title = tag.get_text()
-                rslt.append(AnimeSearchResult(title, id, url))
-    n = len(rslt)
-    desc = "\n".join(f"**#{i+1}** {rslt[i].title}" for i in range(n))
+    rslt = await bot.search_anime(query)
+    if not rslt:
+        return await cmd.send("No matching results found.")
+    desc = "\n".join(f"{choices[i[0]]} {i[1].title}" for i in enumerate(rslt))
     em = discord.Embed(title=f"Search results for {query}", description=desc, color=0x2ECC71)
     msg = await cmd.send(embed=em)
-    for emoji in choices[:n]:
+    for emoji in choices[:len(rslt)]:
         await msg.add_reaction(emoji)
     
 
     def check(reaction, user):
-        return user.id == cmd.author.id and str(reaction) in choices[:n] and reaction.message.id == msg.id
+        return user.id == cmd.author.id and str(reaction) in choices[:len(rslt)] and reaction.message.id == msg.id
 
 
     try:
         reaction, user = await bot.wait_for("reaction_add", check=check, timeout=300.0)
-    except asyncio.TimeOutError:
+    except asyncio.TimeoutError:
         await cmd.send(f"<@!{cmd.author.id}> didn't respond to anime search request. Message timed out.")
         return
     else:
